remove api routes, switch to server functions
